# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DEV-001 Thyroid Pre-Visit Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questionnaires/{appointmentId}/consent:
    post:
      operationId: recordConsent
      tags: [questionnaires]
      summary: Record patient consent (patient auth)
      security:
        - patientAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentInput"
      responses:
        "200":
          description: Consent recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /questionnaires/{appointmentId}:
    get:
      operationId: getQuestionnaire
      tags: [questionnaires]
      summary: Load questionnaire state (patient auth)
      security:
        - patientAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Questionnaire state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionnaireState"
  /questionnaires/{appointmentId}/save:
    post:
      operationId: saveQuestionnaire
      tags: [questionnaires]
      summary: Save questionnaire progress (patient auth)
      security:
        - patientAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionnaireAnswers"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResult"
  /questionnaires/{appointmentId}/submit:
    post:
      operationId: submitQuestionnaire
      tags: [questionnaires]
      summary: Final questionnaire submission (patient auth)
      security:
        - patientAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionnaireAnswers"
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitResult"
  /questionnaires/{appointmentId}/doctor:
    get:
      operationId: getDoctorQuestionnaire
      tags: [questionnaires]
      summary: Clinical questionnaire view (doctor)
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full questionnaire record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionnaireRecord"

  # ── Uploads ───────────────────────────────────────────────────────────────
  /uploads/{appointmentId}:
    post:
      operationId: uploadDocument
      tags: [uploads]
      summary: Upload document metadata (patient auth)
      security:
        - patientAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"

    get:
      operationId: listDocuments
      tags: [uploads]
      summary: List patient's own documents (patient auth)
      security:
        - patientAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
  /uploads/{appointmentId}/status:
    post:
      operationId: updateLabStatus
      tags: [uploads]
      summary: Save the patient's lab/document status
      security:
        - patientAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabStatusInput"
      responses:
        "200":
          description: Lab status saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── Doctor clinical ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    patientAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    StaffUser:
      type: object
      required: [id, email, role, fullName, isActive, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [doctor, clinic_admin, nurse]
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        mfaEnabled:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
...
    Appointment:
      type: object
      required: [id, invitedFullName, invitedPhone, appointmentType, scheduledAt, status, createdAt]
      properties:
        id:
          type: string
        patientId:
          type: ["string", "null"]
        doctorId:
          type: string
        invitedFullName:
          type: string
        invitedPhone:
          type: string
        appointmentType:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft_invitation, link_sent, opened, in_progress, submitted, locked, reopened, rescheduled, cancelled]
        labStatus:
          type: ["string", "null"]
        excludedFromClinicalViews:
          type: boolean
        originalScheduledAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        doctor:
          $ref: "#/components/schemas/StaffUser"
        patient:
          $ref: "#/components/schemas/Patient"
...
    Patient:
      type: object
      required: [id, fullName, phone, dateOfBirth, createdAt]
      properties:
        id:
          type: string
        fullName:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        sex:
          type: ["string", "null"]
        heightCm:
          type: ["string", "null"]
        matchStatus:
          type: string
        duplicateReviewFlag:
          type: boolean
        consentGivenAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    QuestionnaireState:
      type: object
      required: [isLocked, appointment]
      properties:
        questionnaire:
          $ref: "#/components/schemas/QuestionnaireRecord"
        isLocked:
          type: boolean
        appointment:
          $ref: "#/components/schemas/Appointment"
    QuestionnaireRecord:
      type: object
      required: [id, appointmentId, status, schemaVersion, createdAt]
      properties:
        id:
          type: string
        appointmentId:
          type: string
        status:
          type: string
          enum: [not_started, consent_given, in_progress, saved, submitted, locked]
        schemaVersion:
          type: string
        answers:
          type: object
          additionalProperties: true
        consentGivenAt:
          type: ["string", "null"]
          format: date-time
        savedAt:
          type: ["string", "null"]
          format: date-time
        submittedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ConsentInput:
      type: object
      properties:
        consentVersion:
          type: ["string", "null"]
    QuestionnaireAnswers:
      type: object
      required: [answers]
      properties:
        answers:
          type: object
          additionalProperties: true
    SaveResult:
      type: object
      required: [ok, savedAt, questionnaireId]
      properties:
        ok:
          type: boolean
        savedAt:
          type: string
          format: date-time
        questionnaireId:
          type: string
    SubmitResult:
      type: object
      required: [ok, submittedAt]
      properties:
        ok:
          type: boolean
        submittedAt:
          type: string
          format: date-time

    # ── Documents ───────────────────────────────────────────────────────────
    Document:
      type: object
      required: [id, appointmentId, originalFileName, mimeType, fileSizeBytes, createdAt]
      properties:
        id:
          type: string
        appointmentId:
          type: string
        originalFileName:
          type: string
        mimeType:
          type: string
        fileSizeBytes:
          type: integer
        documentType:
          type: ["string", "null"]
        labStatus:
          type: ["string", "null"]
        deletedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    DocumentInput:
      type: object
      required: [originalFileName, mimeType, fileSizeBytes]
      properties:
        originalFileName:
          type: string
        mimeType:
          type: string
        fileSizeBytes:
          type: integer
        documentType:
          type: ["string", "null"]
        labStatus:
          type: ["string", "null"]
    LabStatusInput:
      type: object
      required: [labStatus]
      properties:
        labStatus:
          type: string
          enum:
            - uploaded_digitally
            - will_bring_physical
            - results_pending
            - no_results_available
            - not_required
    UploadResult:
      type: object
      required: [document]
      properties:
        document:
          $ref: "#/components/schemas/Document"
        uploadUrl:
          type: ["string", "null"]
...