# lib/api-zod/src/generated/types/documentInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DEV-001 Thyroid Pre-Visit Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentInput {
  originalFileName: string;
  mimeType: string;
  fileSizeBytes: number;
  /** @nullable */
  documentType?: string | null;
  /** @nullable */
  labStatus?: string | null;
}