# lib/api-zod/src/generated/types/labStatusInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DEV-001 Thyroid Pre-Visit Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LabStatusInputLabStatus } from './labStatusInputLabStatus';

export interface LabStatusInput {
  labStatus: LabStatusInputLabStatus;
}